Interesting problem
So we bound the rgb values to their view-model properties. However, the
code behind is still using the view.Value rather than being notified of
the update in the viewModel. This is a weird situation because the the
Paint method being called from OpenTk. Need to think about this some
more.