GameObjectManager static class
After much thought I have decided that the GameObjectManager class was
duplicating functionality already provided by the GameObject class. The
methods must remain in the GameObject class so I made GameObjectManager
static and decided it was going to be a helper class to do things like
find child game objects in the hierarchy.