Fixed issue opentk truncating shader variable name
https://github.com/opentk/opentk/issues/57

Occassionally, the game would crash with a win32 unhandled error. I
tracked it down to the GL.GetActiveUniform call. It didn't crash in this
method but some time after. If this method wasn't called, the program
ran fine. After a bit of searching, it seems that OpenTK had a variable
name truncating issue.