Added basic height map
Height map is based on indexed drawing tutorial. Adding normals for each
vertex in the up direction makes the surface appear smooth. Also, it is
only 4x4 dimension. Though, I could imagine if it was used for the
ground, it would have so many other objects covering it, the normals
would be irrelevant.